Finally make __del__ work in all cases
