+++ bitcart/utils.py
@@ -1,6 +1,8 @@
 import asyncio
 import inspect
 import json
+import signal
+import threading
 import traceback
 from decimal import Decimal
 from typing import Any, Callable, NoReturn
@@ -85,7 +87,43 @@ async def call_universal(func: Callable, *args: Any, **kwargs: Any) -> Any:
         logger.error(f"Error occured:\n{traceback.format_exc()}")
 
 
+def _get_event_loop() -> asyncio.AbstractEventLoop:
+    try:
+        return asyncio.get_running_loop()
+    except RuntimeError:
+        # NOTE: do NOT remove those 2 lines. Otherwise everything just hangs because of incorrect loop being used
+        if threading.current_thread() is threading.main_thread():
+            return asyncio.get_event_loop_policy().get_event_loop()
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        return loop
+
+
+def get_event_loop() -> asyncio.AbstractEventLoop:
+    """Useful utility for getting event loop. Acts like get_event_loop(), but also creates new event loop if needed
+
+    Returns:
+        asyncio.AbstractEventLoop: event loop
+    """
+    loop = _get_event_loop()
+    if loop.is_closed():
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+    return loop
+
+
 async def idle() -> NoReturn:
     """Useful for making event loop idle in the main thread for other threads to work"""
-    while True:
+
+    def signal_handler(*args, **kwargs):
+        nonlocal is_idling
+
+        is_idling = False
+
+    for s in (signal.SIGINT, signal.SIGTERM, signal.SIGABRT):
+        signal.signal(s, signal_handler)
+
+    is_idling = True
+
+    while is_idling:
         await asyncio.sleep(1)
